telink: tl3238c && tl5218x && tl7218x: workaround for commissioning fail
 - tmp fix commissioning ble fail
